Client

+++ client/src/main/java/org/admin/UI/window/enterpriseWindow/dialog/infos/ClientInfoDialog.java
@@ -22,6 +22,7 @@ import org.admin.model.Client;
 import org.admin.model.Option;
 
 import java.util.List;
+import java.util.Optional;
 
 public class ClientInfoDialog extends Main {
     public static void show(Long clientId, Node node){
@@ -203,7 +204,7 @@ public class ClientInfoDialog extends Main {
             else{errorMsg.setText(response.getMsg());}
         });
         deleteButton.setOnAction(event -> {
-            //TODO ALERT DELETE
+            if(!showDeleteClientConfirmation()) return;
             Response response = DeleteClient.deleteById(token, client.getId());
             if(response.getCode() == 200){
                 dialog.close();
@@ -223,5 +224,23 @@ public class ClientInfoDialog extends Main {
 
         dialog.setScene(dialogScene);
         dialog.showAndWait();
+    }
+
+    public static boolean showDeleteClientConfirmation() {
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Предупреждение");
+        alert.setHeaderText("Удаление клиента");
+        alert.setContentText("Вы уверены, что хотите безвозвратно удалить клиента?");
+
+        // Настраиваем кнопки
+        ButtonType buttonTypeOk = new ButtonType("OK", ButtonBar.ButtonData.OK_DONE);
+        ButtonType buttonTypeCancel = new ButtonType("Отмена", ButtonBar.ButtonData.CANCEL_CLOSE);
+        alert.getButtonTypes().setAll(buttonTypeOk, buttonTypeCancel);
+
+        // Ждём выбора пользователя
+        Optional<ButtonType> result = alert.showAndWait();
+
+        // Возвращаем true, если нажата OK
+        return result.isPresent() && result.get() == buttonTypeOk;
     }
 }
