+++ client/src/main/java/org/admin/UI/window/enterpriseWindow/dialog/infos/OptionInfoDialog.java
@@ -5,10 +5,7 @@ import javafx.geometry.Pos;
 import javafx.geometry.VPos;
 import javafx.scene.Node;
 import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.TextArea;
-import javafx.scene.control.TextField;
+import javafx.scene.control.*;
 import javafx.scene.layout.*;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
@@ -22,6 +19,7 @@ import org.admin.model.Option;
 import org.admin.utils.HelpFuncs;
 
 import java.time.LocalTime;
+import java.util.Optional;
 
 public class OptionInfoDialog extends Main {
     public static void show(Long id, Node node){
@@ -104,7 +102,7 @@ public class OptionInfoDialog extends Main {
         cancelButton.setOnAction(event -> dialog.close());
 
         deleteButton.setOnAction(event -> {
-            //TODO ALERT DELETE
+            if(!showDeleteOptionConfirmation()) return;
             Response response = DeleteOption.deleteById(token, option.getId());
             if(response.getCode() == 200){
                 dialog.close();
@@ -132,5 +130,23 @@ public class OptionInfoDialog extends Main {
 
         dialog.setScene(dialogScene);
         dialog.showAndWait();
+    }
+
+    public static boolean showDeleteOptionConfirmation() {
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Предупреждение");
+        alert.setHeaderText("Удаление услуги");
+        alert.setContentText("Вы уверены что хотите безвозвратно удалить услугу?");
+
+        // Настраиваем кнопки
+        ButtonType buttonTypeOk = new ButtonType("OK", ButtonBar.ButtonData.OK_DONE);
+        ButtonType buttonTypeCancel = new ButtonType("Отмена", ButtonBar.ButtonData.CANCEL_CLOSE);
+        alert.getButtonTypes().setAll(buttonTypeOk, buttonTypeCancel);
+
+        // Ждём выбора пользователя
+        Optional<ButtonType> result = alert.showAndWait();
+
+        // Возвращаем true, если нажата OK
+        return result.isPresent() && result.get() == buttonTypeOk;
     }
 }
